Create Median without sorting.rb
##Problem asked by Dropbox
Create an algorithm to efficiently compute the approximate median of a list of numbers.
More precisely, given an unordered list of N numbers, find an element whose rank is between N / 4 and 3 * N / 4, with a high level of certainty in less than O(N) time.
Constraints: Can't use sort method.

Algorithm: Find the highest and lowest number in the input. Find the number that would lie in between them and find the the number in the array that would come closest to it.